Added support for nested models, fixed various bugs
This is now at a baseline level of functionality as an OpenAPI 3 client.
Requests can be sent and models retrieved, and parameters and data are
accepted.  I'm sure this could be cleaned up and made nicer to use, and
many parts of the spec remain unimplemented, but as a minimum viable
product this appears to work.