Add support for resolving allOf attributes
The allOf construct is described in the spec [here](http://spec.openapis.org/oas/v3.0.3#composition-and-inheritance-polymorphism)
and further [here](https://tools.ietf.org/html/draft-handrews-json-schema-02#section-9.2.1.1).

This change allows for schema composition using the allOf construct.
A schema that defines `allOf` will have all of its attributes merged in
from each schema in the `allOf` in sequence, starting with its default
schema and favoring the schema being merged into it.  For example:

```
components:
  schemas:
    Example:
      type: object
      properties:
        foo:
          type: string
          description: Example property
    Example2:
      allOf:
      - $ref: "#/components/schemas/Example"
      properties:
        foo:
          example: bar
```

Would result in a Schema object with a type of "object" and one
property, "foo" whose type is "string", description is "Example
property" and example is "bar".