Accept "Any" type for examples
[The standard](https://github.com/OAI/OpenAPI-Specification/blob/master/versions/3.1.0.md#schemaObject)
says that example may be "Any" type.  The list given originally was not
exhaustive; for example, if the value of example was parsed as a
datetime by yaml, it would be rejected.  While changing the yaml schema
to not parse strings as datetimes would solve this, I think following
the standard's definition of "any" is more correct, and users of this
library should be able to choose how yaml is configured on their own.