bug: Accept "any" type for Example values
Closes #73

The OpenAPI Standard defines a MediaType.example as [any type](https://github.com/OAI/OpenAPI-Specification/blob/main/versions/3.1.0.md#media-type-object),
and similarly for [Example.value](https://github.com/OAI/OpenAPI-Specification/blob/main/versions/3.1.0.md#example-object).

This library previously did not implement these types in that way,
instead accepting a broad set of types, but not _actually_ any type -
namely, arrays would break since they are handled specially when parsing
the spec.

This change moves to `"*"` as the accepted type, which is handled as the
special-case "any" type when parsing specs.  This results in the
expected behavior.  A test was added to confirm that arrays are accepted
as expected.